将load data命令进行处理的类改为一个 将默认的解析包的处理放入map，以更方便的进行扩展！
